run with file Cotxe.py
exemple of program for module file, this five module are connection with them salfe,
use keys of moving UP, DOWN, RIGHT or LEFT.
For the second car, use other controls of your keyboard, W, S, D, and A, execute (ENTER) on the 
thrieh option, the program import the module 'competicio2.py', and then can play with two players.
With the key 'P', you can stopped the game, if you want replay only puch any key of car's controls.  
The last key 'ESC' only for quit of the game.
surprise in your car's race.